:recycle: Move refresh token consumption to tokenset hook
