+++ frontend/src/hooks/useAuth/index.tsx
@@ -1,15 +1,13 @@
 import { useApolloClient } from "@apollo/client";
 import axios from "axios";
-import { createContext, PropsWithChildren, useCallback, useContext } from "react";
+import { createContext, PropsWithChildren, useContext } from "react";
 import { useNavigate } from "react-router-dom";
 import { RoutePaths } from "src/App";
 import config from "src/config";
 import { useGithubProfile } from "src/hooks/useAuth/useGithubProfile";
 import { useRoles } from "src/hooks/useAuth/useRoles";
 import { useTokenSet } from "src/hooks/useTokenSet";
-import { RefreshToken, TokenSet, User, UserRole } from "src/types";
-
-const TOKEN_VALIDITY_TIME_THRESHOLD = 30;
+import { RefreshToken, User, UserRole } from "src/types";
 
 type AuthContextType = {
   login: (refreshToken: RefreshToken) => void;
@@ -23,37 +21,12 @@ type AuthContextType = {
 
 const AuthContext = createContext<AuthContextType | null>(null);
 
-const checkTokenValidity = (token: TokenSet): boolean => {
-  const creationDate = new Date(token.creationDate);
-  const expirationDate = creationDate.setSeconds(
-    creationDate.getSeconds() + token.accessTokenExpiresIn - TOKEN_VALIDITY_TIME_THRESHOLD
-  );
-  return expirationDate > Date.now();
-};
-
 export const AuthProvider = ({ children }: PropsWithChildren) => {
-  const { tokenSet, setTokenSet, clearTokenSet } = useTokenSet();
+  const { tokenSet, setFromRefreshToken, clearTokenSet } = useTokenSet();
   const navigate = useNavigate();
 
-  const consumeRefreshToken = useCallback(async (refreshToken: RefreshToken) => {
-    const accessToken = await axios.post(`${config.HASURA_AUTH_BASE_URL}/token`, {
-      refreshToken,
-    });
-    if (!accessToken.data) throw new Error("Could not consume refresh token");
-    const newHasuraToken = { ...accessToken.data, creationDate: Date.now() };
-    setTokenSet(newHasuraToken);
-    return newHasuraToken;
-  }, []);
-
-  const getUpToDateHasuraToken = useCallback(async () => {
-    if (tokenSet && !checkTokenValidity(tokenSet)) {
-      return consumeRefreshToken(tokenSet.refreshToken);
-    }
-    return tokenSet;
-  }, [tokenSet]);
-
   const login = async (refreshToken: RefreshToken) => {
-    await consumeRefreshToken(refreshToken);
+    await setFromRefreshToken(refreshToken);
     await client.clearStore();
     navigate(RoutePaths.Profile);
   };
